# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/role":{"post":{"operationId":"createRole","summary":"create role","description":"create role","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"],"security":[{"access-token":[]}]},"get":{"operationId":"findAllRole","summary":"find all role","description":"find all role","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleEntity"}}}}}},"tags":["roles"],"security":[{"access-token":[]}]}},"/api/v1/role/{id}":{"get":{"operationId":"findOneRole","summary":"find one role by id","description":"find one role by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"],"security":[{"access-token":[]}]},"patch":{"operationId":"updateOneRole","summary":"update role by id","description":"update role by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"],"security":[{"access-token":[]}]},"delete":{"operationId":"deleteOneRole","summary":"delete role by id","description":"delete role by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"],"security":[{"access-token":[]}]}},"/api/v1/user":{"post":{"operationId":"createUser","summary":"create user","description":"create user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"],"security":[{"access-token":[]}]},"get":{"operationId":"findAllUser","summary":"find all user","description":"find all user","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"tags":["users"],"security":[{"access-token":[]}]},"patch":{"operationId":"updateOneUser","summary":"update user by id","description":"update user by id","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"],"security":[{"access-token":[]}]}},"/api/v1/user/login":{"post":{"operationId":"login","summary":"access token by email and password","description":"login api for user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["users"],"security":[{"access-token":[]}]}},"/api/v1/user/role/{id}":{"get":{"operationId":"setRole","summary":"access token for assigning role","description":"set role of user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"access-token":[]}]}},"/api/v1/user/{id}":{"get":{"operationId":"findOneUser","summary":"find one user by id","description":"find one user by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"],"security":[{"access-token":[]}]},"patch":{"operationId":"addRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"],"security":[{"access-token":[]}]},"delete":{"operationId":"deleteOneUser","summary":"delete user by id","description":"delete user by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"],"security":[{"access-token":[]}]}},"/api/v1/order":{"post":{"operationId":"createOrder","summary":"create order","description":"create order","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"],"security":[{"access-token":[]}]},"get":{"operationId":"findAllOrder","summary":"find all order","description":"find all order","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderEntity"}}}}}},"tags":["orders"],"security":[{"access-token":[]}]}},"/api/v1/order/page":{"get":{"operationId":"paginationOrder","summary":"order pagination","description":"order pagination","parameters":[{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","example":10,"schema":{"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"type":"array"}},{"name":"filter","required":false,"in":"query","schema":{"type":"array"}},{"name":"search","required":false,"in":"query","schema":{"type":"array"}}],"responses":{"200":{"description":""}},"tags":["orders"],"security":[{"access-token":[]}]}},"/api/v1/order/{id}":{"get":{"operationId":"findOneOrder","summary":"find one order by id","description":"find one order by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"],"security":[{"access-token":[]}]},"patch":{"operationId":"updateOneOrder","summary":"update order by id","description":"update order by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"],"security":[{"access-token":[]}]},"delete":{"operationId":"deleteOneOrder","summary":"delete order by id","description":"delete order by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"],"security":[{"access-token":[]}]}},"/api/v1/label":{"get":{"operationId":"listLabel","summary":"list label","description":"list label","parameters":[],"responses":{"200":{"description":""}},"tags":["labels"],"security":[{"access-token":[]}]}}},"info":{"title":"Operator Hub","description":"App","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"Bearer","bearerFormat":"Token","type":"http","in":"header"}},"schemas":{"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string"},"default":{"type":"boolean","default":false}},"required":["name"]},"RoleEntity":{"type":"object","properties":{"id":{"type":"number"},"deleted":{"format":"date-time","type":"string","default":false},"name":{"type":"string"},"default":{"type":"boolean","default":false}},"required":["id","name"]},"UpdateRoleDto":{"type":"object","properties":{"name":{"type":"string"},"default":{"type":"boolean","default":false}}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["email","username","password"]},"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"deleted":{"format":"date-time","type":"string","default":false},"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["id","email","username","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}}},"RoleDto":{"type":"object","properties":{"roleId":{"type":"number"}},"required":["roleId"]},"CreateOrderDto":{"type":"object","properties":{"productId":{"type":"number"},"userId":{"type":"number"}},"required":["productId"]},"OrderEntity":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"status":{"type":"string"},"orderNumber":{"type":"string"},"assignee":{"type":"number"},"orderTime":{"format":"date-time","type":"string","default":false},"dueTime":{"format":"date-time","type":"string","default":false},"deleted":{"format":"date-time","type":"string","default":false},"productId":{"type":"number"},"userId":{"type":"number"}},"required":["id","name","status","orderNumber","assignee","productId","userId"]},"UpdateOrderDto":{"type":"object","properties":{"productId":{"type":"number"},"userId":{"type":"number"}}}}}}